# Futher optimization of computations on samples with large number of layers