Before:

n_layers = 10,	 time = 4 ms
n_layers = 100,	 time = 25 ms
n_layers = 200,	 time = 58 ms
n_layers = 500,	 time = 214 ms
n_layers = 1000,	 time = 653 ms

After:

n_layers = 10,	 time = 4 ms
n_layers = 100,	 time = 21 ms
n_layers = 200,	 time = 46 ms
n_layers = 500,	 time = 160 ms
n_layers = 1000,	 time = 426 ms

It can be improved further by optimizing the string treatment in `INode` class (see screenshot from kcachegrind attached - mallocs and deletions are also supposedly related to string handling). But as we have discussed a better approach will be just to get rid of parameter pool and related machinery completely. Thus I will leave this optimization until we meet a final decision.

<img width="1920" alt="Screenshot_10000_layers" src="https://user-images.githubusercontent.com/22917299/65508468-9243ee00-ded0-11e9-9c66-a044dd0988dc.png">